moving datamodels to assets
