+++ airflow/api_fastapi/core_api/datamodels/dags.py
@@ -159,18 +159,3 @@ class DAGTagCollectionResponse(BaseModel):
 
     tags: list[str]
     total_entries: int
-
-
-class QueuedEventResponse(BaseModel):
-    """Queued Event serializer for responses.."""
-
-    uri: str
-    dag_id: str
-    created_at: datetime
-
-
-class QueuedEventCollectionResponse(BaseModel):
-    """Queued Event Collection serializer for responses."""
-
-    queued_events: list[QueuedEventResponse]
-    total_entries: int
